whatsapp/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
python-dotenv==1.0.0
